Added AuthenticateComponent (Login Form). Introduced the CanActivateWhenAuthenticatedGuard, which is used in the AppRoutingModule as 'canActivate' service. Added AuthenticationService, which is used in the CanActivateWhenAuthenticatedGuard and the AuthenticateComponent. The AuthenticateComponent and CanActivateWhenAuthenticatedGuard, contain the business logic to return to the specified url after authentication.
Added a logout menu item in the appMenu of the AppComponent and also the corresponding instance method.
Added LDAPService which will be called by the AuthenticationService, but currently not yet implemented.

Added also the ThemePickerModule, which contains a ThemePicker Component and all the logic the pick another theme,
but the switching the angular material theme is not yet implemented.

All default angular material themes are now available in the assets sub directory.
The application currently uses only the indigo-pink.css theme file. For the future I like to add
the functionality to switch from theme.